Move tomcat/undertow/jetty/netty under the 'webserver' package
