+++ agent/agent-plugins/webserver-undertow/src/main/java/org/bithon/agent/plugin/webserver/undertow/interceptor/HttpServerExchangeDispatch.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.undertow.interceptor;
+package org.bithon.agent.plugin.webserver.undertow.interceptor;
 
 import io.undertow.server.HttpServerExchange;
 import io.undertow.servlet.handlers.ServletRequestContext;
