+++ dev/spotbugs-exclude.xml
@@ -87,11 +87,11 @@
 
     <!-- agent-plugin -->
     <Match>
-        <Class name="org.bithon.agent.plugin.jetty.interceptor.HttpChannel$HandleException"/>
+        <Class name="org.bithon.agent.plugin.webserver.jetty.interceptor.HttpChannel$HandleException"/>
         <Bug pattern="NM_CLASS_NOT_EXCEPTION" />
     </Match>
     <Match>
-        <Class name="org.bithon.agent.plugin.tomcat.interceptor.StandardWrapperValve$Exception"/>
+        <Class name="org.bithon.agent.plugin.webserver.tomcat.interceptor.StandardWrapperValve$Exception"/>
         <Bug pattern="NM_CLASS_NOT_EXCEPTION" />
     </Match>
     <Match>
